requirements.txt
openai
python-dotenv
PyPDF2
ipykernel
pandas
openpyxl